+++ checkpoint/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.17] - 2025-06-30
+
 ### Added
 
 - #v1 Add `PreservationPolicy` support in Checkpointer.